Add code to capture errors
Note: There are two different paths that can bring us to an error state.

1. Something fails during the request (or during the transformation of
the data into JSON). e.g. if you replace the URL with gibberish.
2. We successfully make our call to the endpoint and receive data back
but the JSON data defines an error. e.g. if you a change the end_date
in the URL to an invalid date such as '20201529' (there is not 15th
month)